Improve GUI for ARM XOs
Add xo-arm specific extensions.
TODO:
Does not work when we have the extension in device and rename them
because tcz.installed scripts do not run properly
Arrowkeys and game keys are not recognised. Probably needs udev or
a specific xorg.conf entry
Power mamangement is not working, because olpc-kbdshim needs udev
which is not available in armv6/7 TC.